+++ sound.h
@@ -2,6 +2,6 @@
 
 #pragma once
 
-int InitSound(unsigned long freq);
+int InitSound(int freq);
 void FreeSound(void);
 void pop_sample(int l, int r);
